# Pull request - step 1 optimization

Hello. I came to notice that there are a several things that might be better done in data loading process. I am not a big C++ expert, so I recommend a double-check and be available also to discuss these changes in more details. If you tell me on which datasets I should test, I can do that also.

1) The main thing I checked was the time filtering of the point cloud to split it to different worker data. I pre-filter the points by timestamp and then use logarithmic lower/upper bound functions to locate the range of points to use for filtering. The resulting time measurements (of course, very HW and system-dependent but clearly having order of magnitude difference) are below. Note that sorting does not seem to have a major overhead compared to the gained speed at filtering. Faster search algorithm (if timestamps are represented as integer nanoseconds) can be considered. 

* Elapsed time reading: 130874ms (original code)
* Elapsed time reading: 154925ms (with sorting)

* Elapsed time timestamp filtering: 172041ms (original code)
* Elapsed time timestamp filtering: 158ms (applied on sorted points)

I have seen that you have some pretty similar code in the code you commented out. Have you found some situation where proposed solution does not work, or why was it abandoned? 

2) I also commented out writing of `calibrationValidation` file in `lidar_odometry_gui` optional. That makes a huge difference reducing time of reading (lambda function) from 172041ms to 48553ms (including sorting).  The overhead might be small for small files, but once one big laz is used, the overhead is just huge. What if we added some parameter to the interface telling whether to save the calibration validation or not? (I can add it.)

3) There seem to be a bit of a mess with creating `WorkerData`, which always have length of attributes depending on the fixed `threshold` parameter. A lot of memory is constantly cleared and reserved manually for `WorkerData` while this can be dealt with by just changing a bit the whole for-loop and just keeping `WorkerData` inside its scope. The execution of the code dealing with filling of `WorkerData`'s vector then goes from 7307ms to 6265ms. The change is small but consistent between different runs, at least on my computer.

I also noticed that you separate data into different `WorkerData` as if to pass it to multiprocessing, but then just assemble it back to a long vector in e.g. step 2. Is the multiprocessing abandoned? Some issues? No time to implement? Because otherwise I am not sure why to do split-and-merge. It is concerning me as very often I get BSOD at further steps, especially at Consistency. I would happy to check if something can be done there in my free time, if you would be interested.

Best regards,
Valeria