@JanuszBedkowski 2024-09-06
Lidar odometry is crashing on my machine. Please check it on windows.