@iegorval 2024-09-24
> Lidar odometry is crashing on my machine. Please check it on windows.

Hello @JanuszBedkowski,

Thanks for commenting on the pull request.

This was caused by a small bug introduced by me while making the final commit indeed. Realized there was a line of code missing.

But what I also noticed while debugging is this original for cycle you have for filtering:
![image](https://github.com/user-attachments/assets/b5e0f818-a458-4817-b177-7498f56ade6e)
Is it indeed meant to be like this, with that `break` on line 518? If my intuition is correct, you have this `break` just to make the search/filtering stop earlier. But actually this `break` changes the result (on my data at least): more points are filtered out with this `break` that without it, which from condition in line 507 seems like not something that I would expect.

Could you please comment on this one before I make the commit?