@iegorval 2024-09-30
@JanuszBedkowski @michalpelka 

So, I added some changes:

1. I fixed the crash (we tested also on Windows machine). The crash was connected with me accidentally removing the `std::sort` which I still mentioned in the text of the pull request. Incidentally it also showed that the assumption of having all points perfectly sorted by timestamp is incorrect. I also wrote some Python code (as I am better-versed in Python tbh) to check the timestamp filtering and I found out that we throw out ~1% points by assuming that timestamps are aligned (in the original version of the timestamp filtering). As my code does not have this assumption and force-sorts the data, it regains these points. While this is not a huge amount of points, I also do not see reason of getting rid of them. Unless it was by some design choice I am not aware of.
2. While debugging my own problem and cross-checking with Python, I also noticed that there is always one more IMU data row than it has to be. It seems that the file reader picks the empty line on IMU csv file, and while the line is empty, still uses previously-read data to add to the IMU data list. So, while using e.g. IMU 0, we always have at the very end of IMU data, one single row from IMU 1. Again, probably it does not have some big influence, considering the total amount of the IMU poses, but that was probably not done on purpose.